+++ src/artikkelit/generate/generate3xxFields.js
@@ -14,42 +14,3 @@ export function generatef337(isElectronic) {
     return [{code: 'a', value: 'käytettävissä ilman laitetta'}, {code: 'b', value: 'n'}, {code: '2', value: 'rdamedia'}];
   }
 }
-
-// "teatteriarvostelu", "sirkusarvostelu", "elokuva-arvostelu", "tanssiarvostelu", "musiikkiarvostelu" tai "taidearvostelu"
-export function generatef380(reviewType = false) {
-  if (!reviewType) {
-    return [];
-  }
-
-  const subfield2 = {code: '2', value: 'mts/fin'};
-
-  if (reviewType === 'kirja-arvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'kirja-arvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2187'}]}];
-  }
-
-  if (reviewType === 'teatteriarvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'teatteriarvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2180'}]}];
-  }
-
-  if (reviewType === 'sirkusarvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'sirkusarvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2182'}]}];
-  }
-
-  if (reviewType === 'elokuva-arvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'elokuva-arvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2183'}]}];
-  }
-
-  if (reviewType === 'tanssiarvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'tanssiarvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2184'}]}];
-  }
-
-  if (reviewType === 'musiikkiarvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'musiikkiarvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2185'}]}];
-  }
-
-  if (reviewType === 'taidearvostelu') {
-    return [{tag: '380', ind1: ' ', ind2: ' ', subfields: [{code: 'a', value: 'taidearvostelu'}, subfield2, {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:mts:m2186'}]}];
-  }
-
-  return [];
-}
