+++ src/artikkelit/generate/generate6xxFields.js
@@ -172,6 +172,21 @@ function generatef655(vocab, prefLabel, lang, uri, subdivision = false) {
   ];
 }
 
+export function generatef655reviews(reviews) {
+	
+  if (reviews && Array.isArray(reviews) && reviews.length > 0) {
+    return [
+      {tag: '655', ind1: ' ', ind2: '7', subfields:
+        [
+          {code: 'a', value: 'kirja-arvostelut'},
+          {code: '2', value: 'slm/fin'},
+          {code: '0', value: 'http://urn.fi/URN:NBN:fi:au:slm:s1093'}
+        ]}
+    ];
+  }
+  return [];
+}
+
 function selectSubfield(subdivision, code = false) {
   if (subdivision) {
     return [{code, value: subdivision}];
