+++ src/artikkelit/generate/generate6xxFields.spec.js
@@ -1,7 +1,7 @@
 import {READERS} from '@natlibfi/fixura';
 import {expect} from 'chai';
 import generateTests from '@natlibfi/fixugen';
-import {generatef6xxs} from './generate6xxFields.js';
+import {generatef6xxs, generatef655reviews} from './generate6xxFields';
 
 generateTests({
   callback: testF6xx,
@@ -18,6 +18,34 @@ async function testF6xx({getFixture, expectToFail = false}) {
     const input = getFixture('input.json');
     const expectedResults = getFixture('output.json');
     const result = await generatef6xxs(input.ontologyWords);
+
+    expect(result).to.deep.equal(expectedResults);
+    expect(expectToFail, 'This is expected to succes').to.equal(false);
+
+  } catch (error) {
+    if (!expectToFail) {
+      throw error;
+    }
+
+    expect(expectToFail, 'This is expected to fail').to.equal(true);
+  }
+}
+
+generateTests({
+  callback: testF655B,
+  path: [__dirname, '..', '..', '..', 'test-fixtures', 'generatef655B'],
+  useMetadataFile: true,
+  recurse: false,
+  fixura: {
+    reader: READERS.JSON
+  }
+});
+
+async function testF655B({getFixture, expectToFail = false}) {
+  try {
+    const input = getFixture('input.json');
+    const expectedResults = getFixture('output.json');
+    const result = await generatef655reviews(input.reviews);
 
     expect(result).to.deep.equal(expectedResults);
     expect(expectToFail, 'This is expected to succes').to.equal(false);
