--- test-fixtures/generatef380/02/input.json
@@ -1,15 +1,0 @@
-{
-    "article": {
-        "title": "",
-        "titleOther": "",
-        "language": {
-         "iso6391": "fi",
-         "iso6392b": "fin",
-         "ui": "Suomi"
-        },
-        "link": "",
-        "type": "",
-        "reviewType": "musiikkiarvostelu",
-        "sectionOrColumn": ""
-    }
-}
