--- test-fixtures/generatef380/03/input.json
@@ -1,16 +1,0 @@
-{
-"authors": [],
- "article": {
-  "title": "A Modest Proposal",
-  "titleOther": "",
-  "language": {
-   "iso6391": "fi",
-   "iso6392b": "fin",
-   "ui": "Suomi"
-  },
-  "link": "",
-  "type": "",
-  "reviewType": "kirja-arvostelu",
-  "sectionOrColumn": ""
- }
-}
