--- test-fixtures/generatef380/03/output.json
@@ -1,21 +1,0 @@
-[
-    {
-     "tag": "380",
-     "ind1": " ",
-     "ind2": " ",
-     "subfields": [
-      {
-       "code": "a",
-       "value": "kirja-arvostelu"
-      },
-      {
-       "code": "2",
-       "value": "mts/fin"
-      },
-      {
-       "code": "0",
-       "value": "http://urn.fi/URN:NBN:fi:au:mts:m2187"
-      }
-     ]
-    }
-]
