+++ test-fixtures/generatef655B/01/metadata.json
@@ -1,0 +1,9 @@
+{
+  "description": "generatef655 B - reviews / one review",
+  "skip": false,
+  "only": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
