+++ test-fixtures/generatef655B/04/input.json
@@ -1,0 +1,3 @@
+{
+    "reviews": []
+}
