+++ test-fixtures/generatef6xx/17/metadata.json
@@ -1,0 +1,9 @@
+{
+  "description": "generatef6xxx - Field 655 printed when vocab = slm, prefLabel = muistelmat",
+  "skip": false,
+  "only": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
