+++ test-fixtures/generatef6xx/18/input.json
@@ -1,0 +1,12 @@
+{
+    "ontologyWords": [
+        {
+         "key": 1,
+         "prefLabel": "other",
+         "vocab": "otherCategory",
+         "lang": "fi",
+         "uri": "http://urn.fi/URN:NBN:fi:au:slm:s286"
+        }
+       ]    
+}
+
