+++ test-fixtures/generatef6xx/18/metadata.json
@@ -1,0 +1,9 @@
+{
+  "description": "generatef6xxx - Field 655 printed when vocab = otherCategory, prefLabel = other",
+  "skip": false,
+  "only": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
