+++ test-fixtures/generatef6xx/18/output.json
@@ -1,0 +1,13 @@
+[
+    {
+    "tag": "655",
+    "ind1": " ",
+    "ind2": "4",
+    "subfields": [
+        {
+        "code": "a",
+        "value": "other"
+        }    
+    ]        
+    }
+]
